# perf(tutorials): skip some writes in export_vtk_tutorial.py