--- a/.docs/Notebooks/export_vtk_tutorial.py
+++ b/.docs/Notebooks/export_vtk_tutorial.py
@@ -122,13 +122,6 @@ ml.dis.top.export(model_top_dir, fmt="vtk")
 model_bottom_dir = output_dir / "BOTM"
 ml.dis.botm.export(model_bottom_dir, fmt="vtk")
 
-# ### Export transient array recharge
-
-# transient 2d array
-# export recharge
-model_recharge_dir = output_dir / "RECH"
-ml.rch.rech.export(model_recharge_dir, fmt="vtk", pvd=True)
-
 # ### Export HK with point scalars
 #
 
@@ -245,7 +238,8 @@ vtkobj.write(output_dir / "Array_example" / "model.vtu")
 vtkobj = vtk.Vtk(ml, xml=True, pvd=True, vertical_exageration=10)
 
 ## add recharge to the VTK object
-recharge = ml.rch.rech.transient_2ds
+subset = list(range(10))
+recharge = {k: v for k, v in ml.rch.rech.transient_2ds.items() if k in subset}
 vtkobj.add_transient_array(recharge, "recharge", masked_values=[0])
 
 ## write vtk files
@@ -268,8 +262,8 @@ vtkobj = vtk.Vtk(ml, xml=True, pvd=True, vertical_exageration=10)
 spd = ml.wel.stress_period_data
 vtkobj.add_transient_list(spd, masked_values=[0])
 
-## write vtk files
-vtkobj.write(output_dir / "tr_list_example" / "wel_flux.vtu")
+## write vtk files (skipped, slow)
+# vtkobj.write(output_dir / "tr_list_example" / "wel_flux.vtu")
 # -
 
 # ### Adding packages to the `Vtk` object
@@ -320,7 +314,9 @@ hds = HeadFile(head_file)
 # create the vtk object and export heads
 vtkobj = vtk.Vtk(ml, xml=True, pvd=True, vertical_exageration=10)
 vtkobj.add_heads(hds)
-vtkobj.write(workspace / "heads_output_test" / "freyberg_head.vtu")
+
+# skipped, slow
+# vtkobj.write(workspace / "heads_output_test" / "freyberg_head.vtu")
 # -
 
 # ### Export heads as point scalar arrays