Since exporting a transient array looks the same as exporting a normal array, I removed that cell. And commented out 2 more write statements that were each taking ~45s. This brings the runtime from >2 minutes down to ~30 seconds on my machine.

The goal here (eventually) is to get the notebooks to run fast enough for the RTD build service, so we don't have to run them on GH actions and then upload to RTD. Also just for convenience &mdash; it seems a bit daunting and unexpected for a tutorial notebook to spin for minutes. Is there a different transient model we could swap into the VTK export example to avoid commenting out the write statements?

The next slowest are

```
142.58s call     test_notebooks.py::test_notebooks[/Users/wbonelli/dev/flopy/.docs/Notebooks/swi2package_example5.py]
117.58s call     test_notebooks.py::test_notebooks[/Users/wbonelli/dev/flopy/.docs/Notebooks/swi2package_example2.py]
66.68s call     test_notebooks.py::test_notebooks[/Users/wbonelli/dev/flopy/.docs/Notebooks/plot_array_example.py]
65.77s call     test_notebooks.py::test_notebooks[/Users/wbonelli/dev/flopy/.docs/Notebooks/feat_working_stack_examples.py]
63.09s call     test_notebooks.py::test_notebooks[/Users/wbonelli/dev/flopy/.docs/Notebooks/plot_cross_section_example.py]
42.40s call     test_notebooks.py::test_notebooks[/Users/wbonelli/dev/flopy/.docs/Notebooks/raster_intersection_example.py]
```

But this PR does not attempt anything on these yet